brain/requirements.txt
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
anthropic>=0.18.0
flask>=3.0.0
twilio>=9.0.0
